refactor(project): refactor the architecture and the structure of the project
