+++ .gitignore
@@ -37,3 +37,6 @@ lerna-debug.log*
 
 # Envirnment files
 **/.env
+
+# Directory for the project keys
+keys/
