+++ src/application/auth/configs/auth.config.ts
