+++ src/infrastructure/redis/providers/redis.provider.ts
@@ -8,5 +8,5 @@ export interface IRedisProvider {
   getClient(): Redis;
 }
 
-export const REDIS_DB0_PROVIDER = 'redis-db0-provider';
-export const REDIS_DB1_PROVIDER = 'redis-db1-provider';
+export const REDIS_DB0_PROVIDER = 'redis-db0-providers';
+export const REDIS_DB1_PROVIDER = 'redis-db1-providers';
