+++ src/presentation/presentation.module.ts
@@ -1,13 +1,20 @@
 import { Module } from '@nestjs/common';
 import { AuthHttpController } from './http/controllers/auth/auth-http.controller';
 import { ConversationGateway } from '@presentation/ws/gateways/conversation.gateway';
-import { ApplicationModule } from '@application/application.module';
-import { AuthModule } from '@presentation/auth/auth.module';
-import { AuthModule as CoreAuthModule } from '@shared/auth/auth.module';
+import { AuthModule } from '@application/auth/auth.module';
+import { AuthHttpGuard } from '@presentation/guards/auth-http.guard';
+import { AuthWsGuard } from '@presentation/guards/auth-ws.guard';
+import { UserModule } from '@application/user/user.module';
 
 @Module({
-  imports: [ApplicationModule, CoreAuthModule, AuthModule],
+  imports: [AuthModule, UserModule],
   controllers: [AuthHttpController],
-  providers: [ConversationGateway],
+  providers: [
+    // Websocket gateways
+    ConversationGateway,
+    // Guards
+    AuthHttpGuard,
+    AuthWsGuard,
+  ],
 })
 export class PresentationModule {}
