+++ src/presentation/ws/gateways/conversation.gateway.ts
@@ -8,7 +8,7 @@ import {
 } from '@nestjs/websockets';
 import { Logger, UseGuards } from '@nestjs/common';
 import { Server } from 'socket.io';
-import { AuthWsGuard } from '@presentation/auth/guards/auth-ws.guard';
+import { AuthWsGuard } from '@presentation/guards/auth-ws.guard';
 
 @UseGuards(AuthWsGuard)
 @WebSocketGateway({ cors: '*' })
